Created the SCHEMA.md file (with some content)
Created the Schema file which describes basic schema for the entire project.  There are a few things which are in it which probably are part of other sub-projects.  But this is the start of it.

Most of what I have put in there is information about the metadata table, which will be very much a core component.  The problem is, I also wanted to put stuff about all the other tables which that directly refered to.

Which has meant that specifically the *user* table needs to be talked about in there.  Maybe the user table itself is not really part of the "core components".  But the other conneced tables probably are.

Will look at it probably tomorrow.